Add functions for refresh tokens
