--- app/oauth2.py
@@ -1,37 +1,0 @@
-from fastapi import Depends, status, HTTPException
-from jose import JWTError, jwt
-from datetime import datetime, timedelta
-from fastapi.security import OAuth2PasswordBearer
-from sqlalchemy.orm import Session
-from .config import settings
-from . import schemas, database, models
-from .models import User
-
-
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl='login')
-
-SECRET_KEY = settings.SECRET_KEY
-ALGORITHM = settings.ALGORITHM
-ACCESS_TOKEN_EXPIRE_MINUTES = settings.ACCESS_TOKEN_EXPIRE_MINUTES
-
-def create_access_token(data: dict):
-    to_encode = data.copy()
-
-    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-    to_encode.update({"exp": expire})
-
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-
-    return encoded_jwt
-
-def verify_access_token(token: str, credentials_exception):
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        id: str = payload.get("user_id")
-        if id is None:
-            raise credentials_exception
-        token_data = schemas.TokenData(id=id)
-    except JWTError:
-        raise credentials_exception
-    
-    return token_data
